services/app/requirements.txt
# Django web app
django>=5.0
psycopg2-binary
gunicorn
whitenoise
python-decouple
django-tailwind-cli
django-htmx
dj-database-url
djangorestframework>=3.15
djangorestframework-api-key>=3.0

# Chess / analysis
python-chess>=1.999
plotly>=5.20.0
pandas>=2.2.0
anthropic

# Worker pipeline (sync + analysis workers in app/ingest/)
sqlalchemy>=2.0.29
psycopg[binary]>=3.2.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.8.0
requests>=2.32.2
tqdm>=4.66.0
runpod>=1.7.0

# Pinned by Snyk to avoid known vulnerabilities
pyarrow>=14.0.1
pillow>=10.3.0
urllib3>=2.6.3
protobuf>=5.29.6
tornado>=6.5.5
zipp>=3.19.1
pygments>=2.20.0
validators>=0.21.0
numpy>=1.22.2
